Create REALUYOASE3240EXER0101
In the previous class discussion, we programmed a code that predicts the velocity of a free-falling ball at any given time using a for loop control structure. Now, create a program that will similarly predict the velocity of the ball at any given time using a while loop or do while loop structure. In the loop, the computation must still keep on reiterating until the condition is met, i.e., a 0.1% percentage error between the last two computed values is met. The user must still provide the mass, drag coefficient, and desired step size.