Create REALUYOASE3240EXER0102
Write a C program that will calculate and show the sum, difference, product, quotient, and sum of the square roots of two arbitrary numbers given by a user. The numbers can either be positive or negative and are not limited to integer numbers.  All answers should be shown in the specified number of decimal places which is also given by the user.